Align Symbols at Baseline
Align symbols at typesetting baseline (same height and same depth) whenever possible (unless the symbol is an icon). So the symbol glyphs are the same as one would get in the pdf in terms of relative font sizes and vertical alignment.
Meanwhile correct a couple of wrong glyphs or unicodes.